requirements.txt
python-dotenv>=0.19.0
PyJWT>=2.3.0
ttkthemes>=3.2.2
matplotlib>=3.7.1
pillow>=9.5.0
pandas>=2.0.1
argon2-cffi>=23.1.0
